+++ app/src/main/java/org/breezyweather/domain/weather/index/PollenIndex.kt
@@ -30,14 +30,14 @@ enum class PollenIndex(
     @StringRes val pollenName: Int,
     val thresholds: List<Int>,
 ) {
-    ALDER("alder", R.string.pollen_alnus, listOf(0, 10, 50, 100, 300)),
+    ALDER("alder", R.string.pollen_alnus, listOf(0, 10, 60, 100, 500)), // ATMO France
     ASH("ash", R.string.pollen_fraxinus, listOf(0, 30, 100, 200, 400)),
-    BIRCH("birch", R.string.pollen_betula, listOf(0, 10, 50, 100, 300)),
+    BIRCH("birch", R.string.pollen_betula, listOf(0, 10, 60, 100, 500)), // ATMO France
     CHESTNUT("chestnut", R.string.pollen_castanea, listOf(0, 1, 2, 3, 4)), // TODO
     // COTTONWOOD("cottonwood", R.string.pollen_cottonwood, listOf(0, 50, 200, 400, 800)),
     CYPRESS("cypress", R.string.pollen_cupressaceae_taxaceae, listOf(0, 1, 2, 3, 4)), // TODO
     // ELM("elm", R.string.pollen_elm, listOf(0, 30, 50, 100, 200)),
-    GRASS("grass", R.string.pollen_poaeceae, listOf(0, 5, 25, 50, 100)),
+    GRASS("grass", R.string.pollen_poaeceae, listOf(0, 3, 30, 50, 250)), // ATMO France
     HAZEL("hazel", R.string.pollen_corylus, listOf(0, 1, 2, 3, 4)), // TODO
     HORNBEAM("hornbeam", R.string.pollen_carpinus, listOf(0, 1, 2, 3, 4)), // TODO
     // JAPANESE_CYPRESS("cypress", R.string.pollen_japanese_cypress, listOf(0, 3, 11, 19, 39)),
@@ -45,14 +45,14 @@ enum class PollenIndex(
     LINDEN("linden", R.string.pollen_tilia, listOf(0, 1, 2, 3, 4)), // TODO
     // MAPLE("maple", R.string.pollen_maple, listOf(0, 30, 50, 100, 200)),
     MOLD("mold", R.string.pollen_mold, listOf(0, 6500, 13000, 50000, 65000)),
-    MUGWORT("mugwort", R.string.pollen_artemisia, listOf(0, 5, 11, 23, 50)), // TODO: To be checked
+    MUGWORT("mugwort", R.string.pollen_artemisia, listOf(0, 3, 30, 50, 250)), // ATMO France
     OAK("oak", R.string.pollen_quercus, listOf(0, 50, 100, 200, 400)),
-    OLIVE("olive", R.string.pollen_olea, listOf(0, 10, 50, 200, 400)),
+    OLIVE("olive", R.string.pollen_olea, listOf(0, 20, 100, 200, 500)), // ATMO France
     // PINE("pine", R.string.pollen_platanus, listOf(0, 50, 200, 500, 1000)),
     PLANE("plane", R.string.pollen_platanus, listOf(0, 1, 2, 3, 4)), // TODO
     PLANTAIN("plantain", R.string.pollen_plantaginaceae, listOf(0, 1, 2, 3, 4)), // TODO
     POPLAR("poplar", R.string.pollen_populus, listOf(0, 1, 2, 3, 4)), // TODO
-    RAGWEED("ragweed", R.string.pollen_ambrosia, listOf(0, 5, 11, 23, 50)),
+    RAGWEED("ragweed", R.string.pollen_ambrosia, listOf(0, 3, 30, 50, 250)), // ATMO France
     SORREL("sorrel", R.string.pollen_rumex, listOf(0, 1, 2, 3, 4)), // TODO
     TREE("tree", R.string.pollen_tree, listOf(0, 10, 50, 100, 300)),
     URTICACEAE("urticaceae", R.string.pollen_urticaceae, listOf(0, 1, 2, 3, 4)), // TODO